Remove 8/12 second long belt recipes.
https://mods.factorio.com/mod/ZAdventure/discussion/5c20d18682a6f9000d89d7f0
TL;DR: When crafting recipes that require belts like splitters and laboratories, it takes very long because the ZADV recipe is used isntead of the vanilla one. Not sure why these recipes even exist in the first place?
If they should exist please explain to me why and I will try to find another solution to the problem.